move has_args to utils.types, add tests for it
